% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attribute_guidance.R
\name{attribute_guidance}
\alias{attribute_guidance}
\title{Attribute Guidance
The following function uses a lasso regressionn from glmnet in order to select variables and estimate variable contribution to sales.
Based on the coefficients and the number of articles with those attributes, it gives the % of sales driven by each primary attribute, that contributes at least 1% to explian sales.
Note that for the top 3 primary attributes, additional columns are added with the specific figures}
\usage{
attribute_guidance(data_group, primary_attributes, atributte_threshold = 1,
  group, Article_id = "Article_Number", Volume = "norm_sales")
}
\arguments{
\item{data_group}{Data frame that contains the range of a specific grouping, with Article_Number, norm_sales and the attributes in text format}

\item{primary_attributes}{A character vector including the columns of the attributes to be analyzed}

\item{group}{A text that indicates the group analyzed}

\item{Article_id}{A text string indicating the column name of Article_id}

\item{Volume}{A text string indicating the column name of norm_sales}

\item{atributte_theshold}{A number that indicates the minimum contribution accepted for a primary attribute to be displayed, over all sales explained by attributes}
}
\value{
Save a data frame that contains all the information at article level based on the Lasso regression with minimum cross validated error, save primary attribute weights
}
\description{
requires glmnet dummies
}
\examples{
\dontrun{
attribute_guidance<- function(data_group=data_group ,
                             primary_attributes=primary_attributes,
                             atributte_theshold =  1)

}
}
\author{
Alan Fortuny <Alan.FortunySicart@adidas-group.com>
}
