% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabrics_transf.R
\name{fabrics_transf}
\alias{fabrics_transf}
\title{Fabrics Transformation}
\usage{
fabrics_transf(data_group, fabric = "fabric")
}
\arguments{
\item{data_group}{input dataset}

\item{fabric}{A text string indicating the column name of fabric}
}
\value{
a dataset including unique fabric values and the upper limits of each fabric category
}
\description{
It manipulates the fabric variable (which is a character variable) and obtains the upper and lower limit of it. This function also isolates
all possible levels of fabric available in the examined dataset of articles.
}
\examples{
\dontrun{
#Run the function
fabric_limits<-fabrics_transf(data_mart_assort_ARP_clean_v0=data_mart_assort_ARP_clean_v0)
}
}
\author{
Nikolaos Kapodistrias <n.kapodistrias@accenture.com>

Ilektra Kocheila <ilektra.kocheila@accenture.com>
}
