% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytical_settings.R
\name{get_analytical_settings}
\alias{get_analytical_settings}
\title{Analytical setting apparel}
\usage{
get_analytical_settings()
}
\arguments{
\item{saveResults}{A boolean varible set by the user when he wants to save the results or not}

\item{table_sales_attributes}{A table provided by the BUs with the article id, the sales or forecasts, the margin, the BV, the Grouping and some attributes}

\item{sql_txt_v_article_rfd}{describes the columns selected from this table}

\item{output_path}{it specifies the path where the results are going to be stored}

\item{sales_name}{the name of the column sales in the dataset}

\item{price_name}{the name of the column price in the dataset}

\item{article_id_name_sales}{the name of the column article_id in the dataset of BUs}

\item{fabric}{the name of the column fabric in the dataset}

\item{split_by_price}{A boolean value indicating if spliting by price should be performed - receives "YES" and "NO" values}

\item{price_cluster_criteria}{A string, if informed as "Manual", will use the manual price segments, otherwise, automatic segments will be applied}

\item{price_cluster_number}{A number defining the number of preferable clusters to group the price when the criteria is set to "Automatic"}

\item{price_threshold}{A numeric vector indicating the thresholds used to split by price}

\item{split_by_fabric}{A boolean value indicating if spliting by fabric should be performed - receives "YES" and "NO" values}

\item{group_fabric}{a boolean value indicating if you want to group fabric at higher level}

\item{fabric_threshold}{A numeric vector indicating the thresholds used to group the fabric}

\item{atributte_levels}{A numeric value indicating the maximunm level per attribute. If a variable exceeds this number then the rest of its levels are grouped to 'other'.}

\item{atributte_theshold}{A numeric value indicating what is the minimum percentage of importance that we use to select the most important primary attributes on the random forest}

\item{maximum_pvalue}{A number indicating the maximum pvalue accepted on the variables (default value is 0.5)}

\item{max_weight}{A number indicating the maximum of the weight of the primary attributes (calculated as the median impact of secondary levels)}

\item{combi}{A number indicating the number of variable combinations to be tested in the attribute regression (minimum value is 5)}

\item{retain_threshold}{A numeric value indicating the percentage of sales that need to be retained from the deletion of articles}
}
\value{
The analytical settings
}
\description{
The user is able to specify several settings regarding the analysis itself.
}
\examples{

get_analytical_settings()
}
\author{
Nikolaos Kapodistrias <n.kapodistrias@accenture.com>
}
