% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_progress_factory.R
\name{report_progress_factory}
\alias{report_progress_factory}
\title{report_process_factory.R}
\usage{
report_progress_factory(prod_division = c(), severity_level = 3)
}
\arguments{
\item{prod_division}{A string refering to the product division}

\item{severity_level}{An integer specifing the threshold of severity (1 =Error,2=Warning, 3=Info,4=Trace,5=Debug)}
}
\value{
The progress of the code showing messages to the desired severity level
}
\description{
Function that print, according to the severity line
}
\author{
Vadim Zayakin <vadims.zajakins@accenture.com>

Fortuny Sicart, Alan <Alan.FortunySicart@adidas-group.com>
}
